move caption to top level
