+++ config-overrides.js
@@ -1,0 +1,2 @@
+const { override, addBabelPlugins } = require('customize-cra')
+module.exports = override(addBabelPlugins('babel-plugin-styled-components'))
