+++ package.json
@@ -26,9 +26,9 @@
     "yup": "^0.29.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test --env=jsdom",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -48,9 +48,10 @@
   },
   "devDependencies": {
     "@types/big.js": "^6.0.0",
+    "customize-cra": "^1.0.0",
     "node-sass": "^4.14.1",
     "prettier": "2.1.2",
-    "react-app-rewired": "^2.1.6",
+    "react-app-rewired": "^2.1.7",
     "react-scripts": "3.4.3",
     "typescript": "^4.0.3",
     "use-deep-compare-effect": "^1.4.0"
