+++ public/index.html
@@ -5,6 +5,10 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="preload" href="https://conflux-static.oss-cn-beijing.aliyuncs.com/fonts/Circular-Std-Bold.ttf" as="font" crossorigin="anonymous" />
+    <link rel="preload" href="https://conflux-static.oss-cn-beijing.aliyuncs.com/fonts/Circular-Std-Medium.ttf" as="font" crossorigin="anonymous" />
+    <link rel="preload" href="https://conflux-static.oss-cn-beijing.aliyuncs.com/fonts/Circular-Std-Book.ttf" as="font" crossorigin="anonymous" />
+
     <link rel="stylesheet" href="/global.css" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
