+++ src/caption/Caption.module.scss
@@ -1,0 +1,6 @@
+.container {
+  height: 100%;
+  &.large {
+    padding-top: 2rem;
+  }
+}
