+++ src/layout/Main.jsx
@@ -4,7 +4,7 @@ import Modal from '../component/Modal'
 import useConfluxPortal from '../lib/useConfluxPortal'
 import Shuttle from '../shuttle/Shuttle'
 import Token from '../token/Token'
-import Caption from '../token/caption/Caption'
+import Caption from '../caption/Caption'
 import History from '../history/History'
 import Market from '../market/Market'
 import { useTranslation } from 'react-i18next'
